Move pthread locks to own module.
